Use theme translations and fix login-link id (#5)
* Use theme translations and fix login-link id

Switches over to using theme translations instead of theme settings to manage the text content. This just adds a little extra flexibility for admins if they want to translate the text to other languages.

It also replaces the `login-link` id with `cta-login-link` since the `login-link` id is already used in core. The relevant CSS is added along with it.

* Run prettier